Add return arr condition and remove some debugging code
